Implement step calculation logic and internal state management in CoffeeTimerContent and Timeline components
