+++ src/components/Timeline.tsx
@@ -1,16 +1,12 @@
-import React, { useRef, useEffect, useState, useCallback } from 'react';
+import React from 'react';
 import { Box, Typography, useMediaQuery } from '@mui/material';
-import { CalculatedStep, StepStatus } from '../types';
+import { CalculatedStep } from '../types';
 import { useSettings } from '../contexts/SettingsContext';
 
 interface TimelineProps {
   steps: CalculatedStep[];
   currentTime: number;
-  setSteps: React.Dispatch<React.SetStateAction<CalculatedStep[]>>;
-  onTimerComplete: () => void;
   isDence?: boolean;
-  onStepTransition?: (index: number, currentAmount: number, targetAmount: number) => void;
-  onStepStatusChange?: (index: number, oldStatus: StepStatus, newStatus: StepStatus) => void;
 }
 
 const CONTAINER_HEIGHT = 400;
@@ -21,42 +17,16 @@ const SMALL_TIMELINE_WIDTH = 260;
 const STEP_TEXT_MARGIN = 20;
 const FIRST_STEP_OFFSET = 10;
 const FONT_SIZE = '1.1rem';
-const INDICATE_NEXT_STEP_SEC = 5;
 
 const Timeline: React.FC<TimelineProps> = ({
   steps,
-  setSteps,
   currentTime,
-  onTimerComplete,
-  isDence,
-  onStepTransition,
-  onStepStatusChange
+  isDence
 }) => {
   const isSmallScreen = useMediaQuery('(max-width:465px)');
   const { darkMode, language } = useSettings();
-  const totalTime = steps[steps.length - 1]?.timeSec;
+  const totalTime = steps[steps.length - 1]?.timeSec || 1;
   const arrowHeight = isDence ? 12 : 25;
-
-  // ステップの状態を内部で管理し、親コンポーネントへの更新を制御する
-  const [internalSteps, setInternalSteps] = useState<CalculatedStep[]>(steps);
-
-  // 親のstepsが変更されたときだけ内部状態を更新
-  useEffect(() => {
-    setInternalSteps(steps);
-  }, [steps]);
-
-  // 内部的なステップ状態の更新が完了したら親に通知する
-  // throttleを使って更新頻度を制限
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      // 内部ステップと外部ステップが異なる場合のみ更新
-      if (JSON.stringify(internalSteps) !== JSON.stringify(steps)) {
-        setSteps(internalSteps);
-      }
-    }, 100); // 100ms間隔で制限
-
-    return () => clearTimeout(timer);
-  }, [internalSteps, setSteps, steps]);
 
   const formatTime = (seconds: number) => {
     const m = Math.floor(seconds / 60);
@@ -75,62 +45,6 @@ const Timeline: React.FC<TimelineProps> = ({
     const topPos = (time / totalTime) * CONTAINER_HEIGHT;
     return time === 0 ? FIRST_STEP_OFFSET : topPos + FIRST_STEP_OFFSET;
   };
-
-  // ステップのステータス追跡用の参照を追加
-  const previousStepsStatusRef = useRef<Record<number, StepStatus>>({});
-  const previousTimeRef = useRef<number>(currentTime);
-
-  // ステップ状態を更新する関数（内部状態のみを更新）
-  const updateStepStatuses = useCallback((currentTimeValue: number) => {
-    // 前回と同じ時間の場合は更新しない
-    if (previousTimeRef.current === currentTimeValue || internalSteps.length === 0) return;
-    previousTimeRef.current = currentTimeValue;
-
-    const lastStep = internalSteps[internalSteps.length - 1];
-    if (currentTimeValue >= lastStep.timeSec) {
-      onTimerComplete();
-    }
-
-    setInternalSteps(prevSteps => {
-      return prevSteps.map((step, index) => {
-        let newStatus: StepStatus = 'upcoming';
-
-        if (currentTimeValue >= step.timeSec && (index === prevSteps.length - 1 || currentTimeValue < prevSteps[index + 1].timeSec)) {
-          newStatus = 'current';
-        } else if (currentTimeValue >= step.timeSec) {
-          newStatus = 'completed';
-        } else if (currentTimeValue >= step.timeSec - INDICATE_NEXT_STEP_SEC && currentTimeValue < step.timeSec) {
-          newStatus = 'next';
-        }
-
-       // ステータス変更を検出
-       const previousStatus = previousStepsStatusRef.current[index];
-       if (previousStatus !== newStatus) {
-         // ステータス変更を親に通知
-         if (onStepStatusChange) {
-           onStepStatusChange(index, previousStatus || 'upcoming', newStatus);
-         }
-         
-         // 状態が「next」に変わった時だけアニメーション設定
-         if (newStatus === 'next' && onStepTransition) {
-           const currentAmount = index > 0 ? (prevSteps[index - 1].cumulativeMl || 0) : 0;
-           const targetAmount = step.cumulativeMl || 0;
-           onStepTransition(index, currentAmount, targetAmount);
-         }
-
-         // ステータスを保存
-         previousStepsStatusRef.current[index] = newStatus;
-       }
-
-       return { ...step, status: newStatus };
-      });
-    });
-  }, [internalSteps, onStepTransition, onTimerComplete, onStepStatusChange]);
-
-  // Update useEffect for timer
-  useEffect(() => {
-    updateStepStatuses(currentTime);
-  }, [currentTime, updateStepStatuses]);
 
   return (
     <Box
@@ -162,7 +76,7 @@ const Timeline: React.FC<TimelineProps> = ({
           }}
         />
         {/* Render each step using absolute positioning */}
-        {internalSteps.map((step, index) => {
+        {steps.map((step, index) => {
           const topPos = getStepPosition(step.timeSec);
           return (
             <Box
